Refactor: JsonResponseBuilder를 JsonResponse의 내부 클래스로 수정했습니다.
